+++ src/App.js
@@ -4,6 +4,7 @@ import Home from './screens/Home'
 import Camera from './screens/Camera'
 import Speech from './screens/Speech'
 import Games from './screens/Games'
+import Game from './screens/Game'
 import './styles/App.css'
 
 const App = () => (
@@ -13,6 +14,7 @@ const App = () => (
       <Route path="/camera" component={Camera} />
       <Route path="/speech" component={Speech} />
       <Route path="/games" component={Games} />
+      <Route path="/game" component={Game} />
     </Switch>
   </Router>
 )
