+++ src/styles/Games.css
@@ -6,97 +6,42 @@
   height: auto;
 }
 
-#star {
-  position: absolute;
+.game-level-options-container  a {
+  color: rgb(34, 34, 34);
 }
 
-.games-container h2 {
-  text-align: center;
-  margin: 20px 0;
-}
-
-.games-question {
-  color: #fff;
+.games-header {
   padding: 0;
+  text-align: center;
+  color: #fff;
+  margin-top: 30px;
+  margin-bottom: 70px;
 }
 
-.games-options-container {
+.game-level-options-container {
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
+  justify-content: space-around;
   margin: 0 auto;
   width: 75vw;
 }
 
-.games-container .igbo-option {
+.game-level-option-container {
   display: flex;
-  position: relative;
   justify-content: center;
   align-items: center;
-  width: 25vw;
-  height: 16vw;
+  height: 20vh;
+  width: 20vw;
+  background-color: white;
   border-radius: 10px;
-  margin: 25px;
-  background-color: #fff;
+  margin-bottom: 5vw;
   box-shadow: 0px 3px 10px rgba(34, 44, 71, 0.4);
   cursor: pointer;
-  user-select: none;
-  top: 0;
   transition: all .2s;
 }
 
-.games-container .igbo-option:hover {
-  box-shadow: 0px 10px 10px rgba(34, 44, 71, 0.4);
-  top: -10px;
-}
-
-.congratulations-container {
-  color: #fff;
-}
-
-.congratulations-icons {
-  position: absolute;
-  width: 100%;
-  font-size: 9rem;
-  bottom: 20vh;
-}
-
-.post-game-text {
-  position: absolute;
-  color: #fff;
-  width: 100%;
-  bottom: 10vh;
-}
-
-.next-game-options-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  margin: 0 auto;
-}
-
-.next-game-option-container {
-  font-family: 'Avenir', sans-serif;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  height: 13vh;
-  width: 20vw;
-  background-color: #fff;
-  color: rgb(34, 34, 34);
-  border-radius: 10px;
-  margin: 0 5vw;
-  box-shadow: 0px 10px 10px rgba(34, 44, 71, 0.4);
-  bottom: -80vh;
-  opacity: 0;
-  transition: all .2s;
-  cursor: pointer;
-}
-
-.next-game-option-container:hover {
+.game-level-option-container:hover {
   margin-top: -10px;
+  box-shadow: 0px 10px 10px rgba(34, 44, 71, 0.4);
 }
